+++ test/with_dummyserver/test_https.py
@@ -13,14 +13,13 @@ from test import (
     onlyPy26OrOlder, requires_network, TARPIT_HOST
     )
 from urllib3 import HTTPSConnectionPool
-import urllib3.connection
 from urllib3.connection import (
     VerifiedHTTPSConnection,
     UnverifiedHTTPSConnection,
 )
 from urllib3.exceptions import (
-    SSLError, MaxRetryError, ReadTimeoutError, ConnectTimeoutError)
-from urllib3.util import Timeout
+    SSLError, ReadTimeoutError, ConnectTimeoutError)
+from urllib3.util.timeout import Timeout
 
 
 log = logging.getLogger('urllib3.connectionpool')
@@ -205,18 +204,18 @@ class TestHTTPS(HTTPSDummyServerTestCase):
     def test_https_timeout(self):
         timeout = Timeout(connect=0.001)
         https_pool = HTTPSConnectionPool(TARPIT_HOST, self.port,
-                                         timeout=timeout,
+                                         timeout=timeout, retries=False,
                                          cert_reqs='CERT_REQUIRED')
 
         timeout = Timeout(total=None, connect=0.001)
         https_pool = HTTPSConnectionPool(TARPIT_HOST, self.port,
-                                         timeout=timeout,
+                                         timeout=timeout, retries=False,
                                          cert_reqs='CERT_REQUIRED')
         self.assertRaises(ConnectTimeoutError, https_pool.request, 'GET', '/')
 
         timeout = Timeout(read=0.001)
         https_pool = HTTPSConnectionPool(self.host, self.port,
-                                         timeout=timeout,
+                                         timeout=timeout, retries=False,
                                          cert_reqs='CERT_REQUIRED')
         https_pool.ca_certs = DEFAULT_CA
         https_pool.assert_fingerprint = 'CC:45:6A:90:82:F7FF:C0:8218:8e:' \
@@ -262,6 +261,7 @@ class TestHTTPS(HTTPSDummyServerTestCase):
         def new_pool(timeout, cert_reqs='CERT_REQUIRED'):
             https_pool = HTTPSConnectionPool(TARPIT_HOST, self.port,
                                              timeout=timeout,
+                                             retries=False,
                                              cert_reqs=cert_reqs)
             return https_pool